move IncorrectLengthError to separate file and add tests
