+++ pkg/crypto/crypto.go
@@ -518,24 +518,6 @@ func publicKeyFromMontgomery(publicKey PublicKey, sb byte) []byte {
 	return pk
 }
 
-type IncorrectLengthError struct {
-	Name        string
-	Len         int
-	ExpectedLen int
-}
-
-func (e IncorrectLengthError) Error() string {
-	return fmt.Sprintf("incorrect %s length %d, expected %d", e.Name, e.Len, e.ExpectedLen)
-}
-
-func NewIncorrectLengthError(name string, len int, expectedLen int) IncorrectLengthError {
-	return IncorrectLengthError{
-		Name:        name,
-		Len:         len,
-		ExpectedLen: expectedLen,
-	}
-}
-
 func array32FromBase58(s, name string) ([32]byte, error) {
 	var r [32]byte
 	b, err := base58.Decode(s)
